mito-ai: refactor mito toolbar widget
